+++ hazelcast-client/src/test/java/com/hazelcast/client/cluster/ClientClusterServiceMemberListTest.java
@@ -19,8 +19,8 @@ import org.junit.runner.RunWith;
 import java.util.Collection;
 import java.util.Set;
 
-import static com.hazelcast.internal.cluster.memberselector.MemberSelectors.DATA_MEMBER_SELECTOR;
-import static com.hazelcast.internal.cluster.memberselector.MemberSelectors.LITE_MEMBER_SELECTOR;
+import static com.hazelcast.cluster.memberselector.MemberSelectors.DATA_MEMBER_SELECTOR;
+import static com.hazelcast.cluster.memberselector.MemberSelectors.LITE_MEMBER_SELECTOR;
 import static com.hazelcast.test.HazelcastTestSupport.assertTrueEventually;
 import static com.hazelcast.test.HazelcastTestSupport.getNode;
 import static java.util.Arrays.asList;
